Apps/ImageResize/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ks-image-resize"
version = "0.1.0"
description = "Batch image resizing utility with GUI and CLI interfaces"
authors = [
    {name = "Kalponic Studio", email = "info@kalponic.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]
dependencies = [
    "Pillow>=10.0.0",
    "customtkinter>=5.2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "flake8>=4.0.0",
]

[project.scripts]
ks-image-resize = "ks_image_resize.cli.main:main"
ks-image-resize-gui = "ks_image_resize.ui.app:main"

[tool.setuptools]
packages = ["ks_image_resize"]
package-dir = {"" = "src"}